Add files via upload
Added playlist integration and tried to handle errors related to typos. I thought it would be kind of neat to try to implement something like this inside of a gui as it seems like it wouldn't really violate most peoples TOS agreements and it's kind of difficult to get a hold of free streaming content. I tried messing with selenium and kivy at an earlier point and i have kind of abandoned that project anyway. Now it's just kind of a fun learning project. I'm not sure how youtube would react to it seeing as they still get views and people still encounter their adds. Doesn't matter at this point anyway. I will update the source code later.